fix: :truck: Moved block spawn to src/spawn/block.rs
Issue: #39
