+++ src/utils/blockdata.rs
@@ -1,6 +1,5 @@
 use bevy::prelude::*;
 
-pub const BLOCK_COUNT: usize = 4;
 pub const MAX_BLOCKDATA: usize = 4;
 pub const I_BLOCK: [[usize; 16]; 4] = [
     [
